+++ packages/patrol_cli/CHANGELOG.md
@@ -6,6 +6,7 @@
 ## Unreleased
 
 - Add `--record-video` flag to `patrol test` and `patrol develop` to record a video per test case (Android and iOS simulators). (#2741)
+- Fix `patrol test`/`patrol develop` not reading logs from iOS simulators, by streaming the simulator's log via `simctl spawn`. (#3198)
 
 ## 4.6.0